14:07
Fixed a naming issues in createPartialDerivatives where the cluster size was overwritten.

14:22

Fixed a bug in the computation of the partial derivatives where the combination was added instead of being subtracted.

15:38

Fixed a bug in the computation of the partial derivatives where the derivatives were computed in the wrong order and indexed incorrectly. It is now correctly ordered for combination reactions with df_i/dc. 

The concentration variable is not the element that changes at each index, it is the function for which the derivative is calculated. This makes physical sense the concentration of a cluster only changes due to combination reactions if it is non-zero to begin with and combination reactions always decrease the concentration.

16:21

Refactored PSICluster->getDissociationFlux() so that it correctly computes the dissociation for single species clusters.

17:27

Norm of matrix ratio 2.38505e-07 difference 19307.8 (user-defined state)

Good enough for me and within the range that Jed told me to hit.

Fixed a bug in the cubic fit used to calculate the incoming flux where the units were in meters instead of nanometers.

~JJB 20131110 17:29

git-svn-id: file:///Users/MacAir/Dropbox/xolotl-cleaned/trunk/xolotl@200 72105f29-0fe1-43bd-9363-b1e281e962a2